Add entryDecorator
