+++ src/types.ts
@@ -7,7 +7,6 @@ type Project = {
   id: string;
   name: string;
   color: string;
-  enabled: boolean;
 };
 
 type User = {
